+++ src/main/java/de/craftlancer/clstuff/heroes/runnables/PlayerScoreCalculateRunnable.java
@@ -1,14 +1,13 @@
 package de.craftlancer.clstuff.heroes.runnables;
 
+import de.craftlancer.clstuff.heroes.Heroes;
+import de.craftlancer.core.util.Tuple;
+import org.bukkit.scheduler.BukkitRunnable;
+
 import java.util.Comparator;
 import java.util.List;
 import java.util.UUID;
 import java.util.stream.Collectors;
-
-import org.bukkit.scheduler.BukkitRunnable;
-
-import de.craftlancer.clstuff.heroes.Heroes;
-import de.craftlancer.core.util.Tuple;
 
 public class PlayerScoreCalculateRunnable extends BukkitRunnable {
     private Heroes heroes;
@@ -19,9 +18,9 @@ public class PlayerScoreCalculateRunnable extends BukkitRunnable {
     
     @Override
     public void run() {
-        List<Tuple<UUID, Integer>> entries = heroes.getPlugin().getRankings().updateScores().entrySet().stream()
+        List<Tuple<UUID, Double>> entries = heroes.getPlugin().getRankings().updateScores().entrySet().stream()
                 .map(a -> new Tuple<>(a.getKey(), a.getValue().getScore()))
-                .sorted(Comparator.comparingInt(Tuple<UUID, Integer>::getValue).reversed()).limit(3).collect(Collectors.toList());
+                .sorted(Comparator.comparingDouble(Tuple<UUID, Double>::getValue).reversed()).limit(3).collect(Collectors.toList());
         
         new PlayerScoreApplyRunnable(entries, heroes).runTask(heroes.getPlugin());
     }
